Fix compile errors ereferrals test by dhis2 dependencies: - I have moved dhis dependant test under cnm variant
