+++ app/src/ereferrals/java/org/eyeseetea/malariacare/data/database/utils/ExportDataStrategy.java
@@ -1,0 +1,4 @@
+package org.eyeseetea.malariacare.data.database.utils;
+
+public class ExportDataStrategy extends AExportDataStrategy {
+}
