--- app/src/ereferrals/java/org/eyeseetea/malariacare/database/utils/ExportDataStrategy.java
@@ -1,6 +1,0 @@
-package org.eyeseetea.malariacare.database.utils;
-
-import org.eyeseetea.malariacare.data.database.utils.AExportDataStrategy;
-
-public class ExportDataStrategy extends AExportDataStrategy {
-}
