+++ app/src/testereferrals/java/org/eyeseetea/malariacare/domain/entity/SettingsTest.java
